requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.2
python-decouple>=3.8
channels>=4.0
daphne>=4.0
python-dateutil>=2.8
PyMuPDF>=1.23.0
pdf2image>=1.17.0
openai>=1.0.0
anthropic>=0.18.0
jira>=3.5.0
openpyxl>=3.1.0
